added banner, added discover habit template, refactoring css code for habit templates
