+++ src/components/habits/habit - get/GetHabitsTemplate.js
@@ -16,46 +16,48 @@ const GetHabitsTemplate = ({ habitColor, habitLogo, habit }) => {
   const { auth } = useContext(AuthContext);
 
   return (
-    <HabitColorStyle>
-      {/*----- DESKTOP ------*/}
-      <DesktopWHeight>
-        <div className={habitColor}>
-          <Element>
-            <span className={`material-symbols-outlined ${habitColor}`}>
-              {habitLogo}
-            </span>
-            <h4>{habit.name}</h4>
-
-            {auth.user.role === "ROLE_USER" && (
-              <div className="complete-btn">
-                <CompleteHabit habit={habit} />
-              </div>
-            )}
-          </Element>
-        </div>
-      </DesktopWHeight>
-
-      {/*----- MOBILE ------*/}
-      <MobileWHeight>
-        <ContainerMobile>
+    <>
+      <HabitColorStyle>
+        {/*----- DESKTOP ------*/}
+        <DesktopWHeight>
           <div className={habitColor}>
-            <HabitMobile>
+            <Element>
+              <span className={`material-symbols-outlined ${habitColor}`}>
+                {habitLogo}
+              </span>
+              <h4>{habit.name}</h4>
+
               {auth.user.role === "ROLE_USER" && (
                 <div className="complete-btn">
                   <CompleteHabit habit={habit} />
                 </div>
               )}
-              <Element>
-                <span className={`material-symbols-outlined ${habitColor}`}>
-                  {habitLogo}
-                </span>
-                <h4>{habit.name}</h4>
-              </Element>
-            </HabitMobile>
+            </Element>
           </div>
-        </ContainerMobile>
-      </MobileWHeight>
-    </HabitColorStyle>
+        </DesktopWHeight>
+
+        {/*----- MOBILE ------*/}
+        <MobileWHeight>
+          <ContainerMobile>
+            <div className={habitColor}>
+              <HabitMobile>
+                {auth.user.role === "ROLE_USER" && (
+                  <div className="complete-btn">
+                    <CompleteHabit habit={habit} />
+                  </div>
+                )}
+                <Element>
+                  <span className={`material-symbols-outlined ${habitColor}`}>
+                    {habitLogo}
+                  </span>
+                  <h4>{habit.name}</h4>
+                </Element>
+              </HabitMobile>
+            </div>
+          </ContainerMobile>
+        </MobileWHeight>
+      </HabitColorStyle>
+    </>
   );
 };
