+++ src/pages/discover/Discover.styles.js
@@ -10,7 +10,18 @@ export const DiscoverStyles = styled.div`
     margin: 0;
     display: flex;
     flex-direction: column;
-    gap: var(--gap-l);
+
+    .filter-container {
+      display: flex;
+      justify-content: flex-end;
+      margin: var(--margin-xs) var(--margin-s) 0 var(--margin-s);
+    }
+
+    .filter-button {
+      display: flex;
+      align-items: center;
+      gap: var(--gap-s);
+    }
   }
 `;
 
@@ -32,82 +43,11 @@ export const Filter = styled.div`
 
   h1 {
     text-align: center;
-    border-bottom: 1px solid var(--color-white-background);
+    border-bottom: var(--border-width-xxs) solid var(--color-white-background);
     font-weight: lighter;
   }
 
   @media (max-width: 1024px) {
     padding: 0;
   }
-`;
-
-export const DiscoverMobile = styled.div`
-  .filter-container {
-    display: flex;
-    justify-content: flex-end;
-    margin: var(--margin-xs) var(--margin-s) var(--margin-s) var(--margin-s);
-  }
-
-  .filter-button {
-    display: flex;
-    align-items: center;
-    gap: var(--gap-s);
-  }
-`;
-
-export const Banner = styled.div`
-  width: 100%;
-  height: 30vh;
-  background-color: var(--color-orange-active);
-  border-radius: var(--border-radius-m);
-  position: relative;
-  margin-top: var(--margin-s);
-
-  img {
-    height: 100%;
-    width: 100%;
-    object-fit: cover;
-    border-radius: var(--border-radius-m);
-    box-shadow: var(--shadow-medium);
-  }
-
-  .banner-text {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 20%;
-    display: flex;
-    flex-direction: column;
-    gap: var(--gap-m);
-    margin: var(--margin-l);
-
-    h1 {
-      color: var(--color-orange);
-    }
-  }
-
-  @media (max-width: 1024px) {
-    min-height: 30vh;
-    margin-top: 0;
-    text-align: center;
-
-    .banner-text {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      width: 100%;
-      margin: var(--margin-xs);
-    }
-
-    p {
-      width: 80%;
-    }
-  }
-`;
-
-export const Habits = styled.div`
-  margin-top: var(--margin-s);
-  display: flex;
-  gap: var(--gap-s);
-  overflow-x: scroll;
 `;
